DecompUtil

Documentation for DecompUtil.

DecompUtil.Hyperplane

DecompUtil.LinearConstraint

DecompUtil.LinearConstraints

DecompUtil.LinearConstraints

DecompUtil.Polyhedron

DecompUtil.seedDecomp

DecompUtil.seedDecomp_2D

DecompUtil.seedDecomp_3D

DecompUtil.Hyperplane — Type

Hyperplane{N, F}(p, n)

represents a hyperplane passing through the point p with normal vector n. N is the dimension, F is the type of the vector. Uses StaticVectors to represent p and n internally.

DecompUtil.LinearConstraint — Type

LinearConstraint(a, b)

represents the constraint $a^T x \leq b$

DecompUtil.LinearConstraints — Type

LinearConstraints(A, b)

represents the linear constraints $A x \leq b$

DecompUtil.LinearConstraints — Method

LinearConstraints(P::Polyhedron)

constructs a linear constraints matrix from a polyhedron

DecompUtil.Polyhedron — Type

Polyhedron(vs::Vector{Hyperplane})

represents a 2D polyhedron using a set of hyperplanes. Each hyperplane's normal points outside the set.

DecompUtil.seedDecomp — Function

seedDecomp(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 2D or 3D space.

Inputs:

pos is the starting point of the seed

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1], [x2, y2], ....]. You can pass in a vector of static arrays if you want.

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the library is called again with a larger max_poly.

Returns:

Polyhedron representing the free space.

DecompUtil.seedDecomp_2D — Function

seedDecomp_2D(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 2D space.

Inputs:

pos is the starting point of the seed

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1], [x2, y2], ....]

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the library is called again with a larger max_poly.

Returns:

Polyhedron representing the free space.

DecompUtil.seedDecomp_3D — Function

seedDecomp_3D(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 3D space.

Inputs:

pos is the starting point of the seed, e.g. [x0, y0, z0]

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1, z1], [x2, y2, z2], ....]

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the function is called again with a larger max_poly.

Returns:

Polyhedron representing the free space.