DecompUtil

Documentation for DecompUtil.

This is a julia wrapper to DecompUtil, created by Devansh Agrawal. I have added a few small functionalities, but have not exported all of the functions provided by the original authors. If there are some particular functions you would like to be able to use, let me know and I should be able to wrap them without much difficulty.

This is research code, and so there is no guarantee of correctness, but so far it has worked well for me. If you find an error, please let me know by raising a Github Issue.

All credits should go to the original authors:

@article{liu2017planning,
  title={Planning dynamically feasible trajectories for quadrotors using safe flight corridors in 3-d complex environments},
  author={Liu, Sikang and Watterson, Michael and Mohta, Kartik and Sun, Ke and Bhattacharya, Subhrajit and Taylor, Camillo J and Kumar, Vijay},
  journal={IEEE Robotics and Automation Letters},
  volume={2},
  number={3},
  pages={1688--1695},
  year={2017},
  publisher={IEEE}
}

For those interested, the wrapper works by exposing a C interface to the original library, creating a JLL package using BinaryBuilder, and then creating the Julia package.

Quick Start

The main export of this library is seedDecomp. This performs a decomposition in a 2D or 3D space, and returns a set of hyperplanes that indicate the safe flight corridor.

2D example:

using DecompUtil

# seed from which the decomposition starts
pos = [0,0.]

# define all the obstacle points
obs = [[-0.2, 1.5], [1, 0.], [0.8,-1.], [ -0.5, -0.5]]

# defines the bounding box (i.e., -2:2 on x axis, -2:2 on y axis)
bbox = [2,2.]

# I am honestly not too sure what this does,
# and it doesnt seem to have an impact on the result.
# (but it needs to be greater than 0)
dilation_radius = 0.1

result = seedDecomp(pos, obs, bbox, dilation_radius)

8-element Vector{DecompUtil.Hyperplane{2, Float32}}:
 DecompUtil.Hyperplane{2, Float32}(Float32[-0.5, -0.5], Float32[-0.70710677, -0.70710677])
 DecompUtil.Hyperplane{2, Float32}(Float32[1.0, 0.0], Float32[1.0, 0.0])
 DecompUtil.Hyperplane{2, Float32}(Float32[0.8, -1.0], Float32[0.62469506, -0.7808688])
 DecompUtil.Hyperplane{2, Float32}(Float32[-0.2, 1.5], Float32[-0.13216372, 0.9912279])
 DecompUtil.Hyperplane{2, Float32}(Float32[0.0, 2.0], Float32[0.0, 1.0])
 DecompUtil.Hyperplane{2, Float32}(Float32[0.0, -2.0], Float32[-0.0, -1.0])
 DecompUtil.Hyperplane{2, Float32}(Float32[2.0, 0.0], Float32[1.0, 0.0])
 DecompUtil.Hyperplane{2, Float32}(Float32[-2.0, 0.0], Float32[-1.0, -0.0])

This returns a vector of Hyperplanes. Each hyperplane v has a point v.p and a normal v.n and defines the free area as the halfspace:

\[\{ x : n^T x \leq n^T p \}\]

We can convert it to a constraints matrix of the form $Ax \leq b$:

A, b = constraints_matrix(result)

A

8×2 Matrix{Float32}:
 -0.707107  -0.707107
  1.0        0.0
  0.624695  -0.780869
 -0.132164   0.991228
  0.0        1.0
 -0.0       -1.0
  1.0        0.0
 -1.0       -0.0

b

8-element Vector{Float32}:
 0.70710677
 1.0
 1.2806249
 1.5132747
 2.0
 2.0
 2.0
 2.0

and we can verify that the seed lies inside this polyhedron:

A * pos - b

8-element Vector{Float64}:
 -0.7071067690849304
 -1.0
 -1.2806248664855957
 -1.5132746696472168
 -2.0
 -2.0
 -2.0
 -2.0

A * pos <= b

true

Since we provide a bounding box, the result is always a bounded polyhedron.

We can visualize it as follows:

using Plots

function plot_2D!(pos, obs, bbox, res)

    scatter!([pos[1]], [pos[2]], label="Seed")
    scatter!(first.(obs), last.(obs), label="Obstacles")
    vline!([pos[1] - bbox[1], pos[1] + bbox[1]], label="Bounding Box", linestyle=:dash, linewidth=2, color=:black)
    hline!([pos[2] - bbox[2], pos[2] + bbox[2]], label=false, linestyle=:dash, linewidth=2, color=:black)

    for (i, v) in enumerate(res)
        p = v.p
        n = v.n
        lab = i==1 ? "Hyperplane" : false
        plot!(t -> p[1] + n[2]*t, t-> p[2] - n[1]*t, -4,4, label=lab, color=:blue)
    end

    # plot!(legend = :outerleft)

end

plot()
plot_2D!(pos, obs, bbox, result)
xlims!(-2.5, 2.5)
ylims!(-2.5, 2.5)
plot!(aspect_ratio=:equal)

Finally, for convenience, a shrink function is provided, that shrinks the polyhedron by a distance d

shrunk_result = DecompUtil.shrink(result, 0.3)

plot()
plot_2D!(pos, obs, bbox, shrunk_result)
xlims!(-2.5, 2.5)
ylims!(-2.5, 2.5)
plot!(aspect_ratio=:equal)

Note, there is no guarantee that the shrunk polyhedron contains the seed, or indeed is non-empty. Also, note the polyhedron can have redundant constraints.

Note, no plotting utilities are not exported, so as to not introduce dependencies on plotting.

Reference

DecompUtil.Hyperplane

DecompUtil.constraints_matrix

DecompUtil.seedDecomp

DecompUtil.seedDecomp_2D

DecompUtil.seedDecomp_3D

DecompUtil.shrink

DecompUtil.Hyperplane — Type

Hyperplane{N, F}(p, n)

represents a hyperplane passing through the point p with normal vector n. N is the dimension, F is the type of the vector. Uses StaticVectors to represent p and n internally.

Example: Hyperplane([1.0, 0.0], [1.0, 1.0])

DecompUtil.constraints_matrix — Method

constraints_matrix(vs::Vector{Hyperplane})

constructs a linear constraints matrix from a list of hyperplanes

DecompUtil.seedDecomp — Function

seedDecomp(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 2D or 3D space.

Inputs:

pos is the starting point of the seed

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1], [x2, y2], ....]. You can pass in a vector of static arrays if you want.

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the library is called again with a larger max_poly.

Returns:

Polyhedron representing the free space.

DecompUtil.seedDecomp_2D — Function

seedDecomp_2D(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 2D space.

Inputs:

pos is the starting point of the seed

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1], [x2, y2], ....]

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the library is called again with a larger max_poly.

Returns:

Polyhedron representing the free space.

DecompUtil.seedDecomp_3D — Function

seedDecomp_3D(pos, obs, bbox, dilation_radius, max_poly=1)

Perform a seed decomposition in a 3D space.

Inputs:

pos is the starting point of the seed, e.g. [x0, y0, z0]

obs is a vector of points representing obstacles in the environment. In the format [[x1, y1, z1], [x2, y2, z2], ....]

bbox is the size of the bounding box within which the decomposition should happen

dilation_radius is the dilation radius (refer to original paper)

max_poly is the assumed maximum number of hyperplanes in the resulting solution. If this is smaller than the true number, the function is called again with a larger max_poly.

Returns:

Vector{Hyperplane} representing the free space.

DecompUtil.shrink — Method

shrink(vs::Vector{Hyperplane}, d)

returns a new list of hyperplanes, where the free space has been decreased by a margin d. You can pass in negative values for d if you want to expand the region. There is no guarantee that the resulting polyhedron is non-empty, or that it doesnt include redundant constraints.